+++ packages/contracts/src/support/meta/MetaTargetContext.ts
@@ -1,3 +1,5 @@
+import { Context } from './types';
+
 /**
  * Meta Decorator Target Context
  */
@@ -5,16 +7,29 @@ export default interface MetaTargetContext
 {
     /**
      * The class that owns the meta
+     * 
+     * @type {object}
      */
-    owner: object,
+    owner: object;
 
     /**
      * "This" argument
+     *
+     * @type {any}
      */
-    thisArg: any, /* eslint-disable-line @typescript-eslint/no-explicit-any */
+    thisArg: any; /* eslint-disable-line @typescript-eslint/no-explicit-any */
 
     /**
      * The target class, field, method... that is being decorated
+     *
+     * @type {object}
      */
-    target: object,
+    target: object;
+
+    /**
+     * Decorator context
+     *
+     * @type {Context}
+     */
+    context: Context;
 }