+++ packages/contracts/src/support/meta/types.ts
@@ -1,28 +1,10 @@
-import MethodContext from "./MethodContext";
-import GetterContext from "./GetterContext";
-import SetterContext from "./SetterContext";
-import FieldContext from "./FieldContext";
-import AccessorContext from "./AccessorContext";
-import MetaEntry from "./MetaEntry";
 import type { Key } from "@aedart/contracts/support";
+import MetaEntry from "./MetaEntry";
 
 /**
  * Decorator context types for any decorator
  */
 export type Context = DecoratorContext;
-
-/**
- * @deprecated Replaced by {@link ClassMemberDecoratorContext}
- * 
- * Decorator context types for class element decorators
- */
-export type MemberContext =
-    | MethodContext
-    | GetterContext
-    | SetterContext
-    | FieldContext
-    | AccessorContext
-    ;
 
 /**
  * Callback that returns a meta entry object.
@@ -38,6 +20,15 @@ export type MetadataRecord = DecoratorMetadata;
  * Reference to the owner object that contains metadata 
  */
 export type MetaOwnerReference = WeakRef<object>;
+
+/**
+ * Initializer callback
+ * 
+ * @see ClassDecoratorContext.addInitializer
+ */
+export type InitializerCallback = (
+    this: any /* eslint-disable-line @typescript-eslint/no-explicit-any */
+) => void;
 
 /**
  * A location (key or path) to a metadata entry, in a given owner object