+++ packages/contracts/src/types.ts
@@ -1,3 +1,5 @@
+export * from './decorators';
+
 /**
  * Primitive value
  *