+++ packages/support/src/meta/exceptions/index.ts
@@ -1,0 +1,4 @@
+import MetaError from "./MetaError";
+export {
+    MetaError
+}