+++ packages/support/src/meta/index.ts
@@ -1,2 +1,16 @@
+import Entry from "./Entry";
+import MetaRepository from "./MetaRepository";
+import TargetContext from "./TargetContext";
+export {
+    Entry,
+    MetaRepository,
+    TargetContext,
+}
+
+export * from './getMetaRepository';
+export * from './hasMeta';
+export * from './getAllMeta';
+export * from './getMeta';
 export * from './meta';
-export * from './targetMeta';
+export * from './exceptions';
+export * from './target/index'