+++ packages/support/src/meta/target/getTargetMetaRepository.ts
@@ -1,0 +1,12 @@
+import type { TargetRepository } from "@aedart/contracts/support/meta";
+import TargetMetaRepository from "./TargetMetaRepository";
+
+/**
+ * Returns a new Target Meta Repository
+ * 
+ * @return {TargetRepository}
+ */
+export function getTargetMetaRepository(): TargetRepository
+{
+    return TargetMetaRepository.make();
+}